introduce envs package; remove state for now
