+++ clemcore/playpen/__init__.py
@@ -1,6 +1,6 @@
 from clemcore.playpen.callbacks import BaseCallback, GameRecordCallback, RolloutProgressCallback, CallbackList
 from clemcore.playpen.base import BasePlayPen
-from clemcore.playpen.env import GameEnv, make_env
+from clemcore.playpen.envs.env import GameEnv, make_env
 
 __all__ = [
     "BaseCallback",
