+++ clemcore/playpen/base.py
@@ -22,11 +22,11 @@ class BasePlayPen(abc.ABC):
         self.callbacks.on_rollout_start(game_env, self.num_timesteps)
         num_rollout_steps = 0
         while num_rollout_steps < rollout_steps:
-            player, context, state = game_env.observe()
+            player, context = game_env.observe()
             response = player(context)
             done, info = game_env.step(response)
             if self.is_learner(player):
-                self.rollout_buffer.append((state, context, response, done, info))
+                self.rollout_buffer.append((context, response, done, info))
                 num_rollout_steps += 1
                 self.num_timesteps += 1
             self.callbacks.update_locals(locals())
